requirements.txt
google-generativeai
pandas
streamlit
pinecone-client
sentence-transformers
python-dotenv
langchain
langchain-google-genai
streamlit-chat